♻️ vue-dot: Refactor typings
