+++ packages/vue-dot/src/patterns/FileList/types.d.ts
@@ -1,0 +1,6 @@
+export interface FileItem {
+	id: string;
+	title: string;
+	state: string;
+	name?: string;
+}
