+++ packages/vue-dot/src/patterns/FileUpload/errorCodes.ts
@@ -1,0 +1,6 @@
+/** Standardized error codes */
+export enum ErrorCodes {
+	MULTIPLE_FILES_SELECTED = 'MULTIPLE_FILES_SELECTED',
+	FILE_TOO_LARGE = 'FILE_TOO_LARGE',
+	FILE_EXT_NOT_ALLOWED = 'FILE_EXT_NOT_ALLOWED'
+}
