+++ packages/vue-dot/src/patterns/FileUpload/types.d.ts
@@ -1,0 +1,4 @@
+export interface ErrorEvent {
+	file: File | DataTransferItemList;
+	code: string;
+}
