+++ packages/vue-dot/src/patterns/LangBtn/types.d.ts
@@ -1,0 +1,10 @@
+interface Language {
+	direction: string;
+	name: string;
+	nativeName: string;
+}
+
+export interface Languages {
+	// 'key' is the code of the language
+	[key: string]: Language;
+}
