+++ packages/vue-dot/src/patterns/PaginatedTable/types.d.ts
@@ -1,0 +1,11 @@
+// Vuetify's VDataTable options
+export interface Options {
+	page: number;
+	itemsPerPage: number;
+	sortBy: string[];
+	sortDesc: boolean[];
+	groupBy: string[];
+	groupDesc: boolean[];
+	multiSort: boolean;
+	mustSort: boolean;
+}
