+++ packages/vue-dot/src/patterns/UploadWorkflow/types.d.ts
@@ -1,0 +1,12 @@
+export interface FileListItem {
+	id: string;
+	title: string;
+	state?: string;
+	name?: string;
+	file?: File;
+}
+
+export interface SelectItem {
+	text: string;
+	value: string;
+}
