+++ packages/vue-dot/src/types.d.ts
@@ -1,0 +1,4 @@
+import Vue from 'vue';
+
+// Export generic type
+export type Refs<T extends object> = Vue['$refs'] & T;
