+++ packages/vue-dot/types/index.d.ts
@@ -21,6 +21,3 @@ export default VueDot;
 export interface VueDotPlugin {
 	install: PluginFunction<VueDotOptions>;
 }
-
-// Export generic type
-export type Refs<T extends object> = Vue['$refs'] & T;
